NetworkPkg/NvmeOfDxe: Populate NBFT table for failure cases.
This change is to populate the details of failed connections to NBFT.
This applies to such subsystems, for which attempt has been configured
but the NVMe-oF driver was unable to create a successful connection to
them. This change will enable the booting OS to get the details of such
subsystems from NBFT and retry the connection to proceed with booting
from them.

Signed-off-by: Amit Jain <amit_jain9@dell.com>
Co-authored-by: Trevor Cockrell <trevor_cockrell@dell.com>