NetworkPkg/NvmeOfDxe: Fix SSNS Primary HFI Descriptor Index assignment
HfiHeader was always pointing to a head of the HFI Descriptor List
and its index was assigned to a SSNS Primary HFI Descriptor Index,
effectively always filling it with a value of '1'. The real HFI
index to be used was added only as a first Secondary HFI
Descriptor Index.

In a multipath scenario where separate SSNS records were defined
for every NIC/HFI, this led to connection failures due
to mismatched network subnets.

Signed-off-by: Tomas Bzatek <tbzatek@redhat.com>